+++ apps/fleetplanner-web/src/App.tsx
@@ -13,6 +13,7 @@ import { FeedbackPage } from "./pages/FeedbackPage";
 import { TemplatesPage } from "./pages/TemplatesPage";
 import { RoadmapPage } from "./pages/RoadmapPage";
 import { GuildSettingsPage } from "./pages/GuildSettingsPage";
+import { PartnershipsPage } from "./pages/PartnershipsPage";
 import { LoginPage } from "./pages/LoginPage";
 import { ErrorState } from "./components/ErrorState";
 import { Avatar } from "./components/Avatar";
@@ -189,6 +190,7 @@ export function App() {
           <Route path="/feedback" element={<FeedbackPage session={session} />} />
           <Route path="/roadmap" element={<RoadmapPage />} />
           <Route path="/guilds/settings" element={<GuildSettingsPage session={session} />} />
+          <Route path="/guilds/partnerships" element={<PartnershipsPage session={session} />} />
           <Route path="/ops/:id/edit" element={<EditOpPage session={session} />} />
           <Route path="/ops/:id" element={<OpDetailPage session={session} />} />
           <Route path="/login" element={<LoginPage />} />